% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_sim.R
\name{demand_sim}
\alias{demand_sim}
\title{Simulate the sales for the introduction and growth phases of new products}
\usage{
demand_sim(
  products_number,
  periods_number,
  shape_number,
  level_number,
  level_range = 1000:10000,
  noise_cv
)
}
\arguments{
\item{products_number}{Number of products}

\item{periods_number}{Number of periods of the introduction and growth phases}

\item{shape_number}{Number of generic shapes}

\item{level_number}{Number of generic levels}

\item{level_range}{Range of values from which the level is sampled}

\item{noise_cv}{The coefficient of variation of the noise added to the simulated sales}
}
\value{
A date frame of three colunms product_id, shape and assigned_shape,
         level and assigned_level, sale, noise and modified_sale - Max between sale+noise and 0
}
\description{
Simulate the sales for the introduction and growth phases of new products
}
\examples{
demand_sim(products_number=100,periods_number=20,shape_number=5,level_number=20,
level_range=1000:10000,noise_cv=0.05)
}
